simplify right panel
